+++ CHANGELOG.md
@@ -43,6 +43,8 @@ Below, you can find a list of changes, comparing with legacy ManeTools.
 - `Scene management` hotkeys moved under the `File` menu with improved selected scene detection.
 - `GetRequiredComponent` is `GetOrAddComponent` now, for clearance.
 - `ArrayElements` attribute renamed to `CollectionItemName`.
+- `DropdownList` attribute now works not only with public methods, but also with privates and properties.
+- `DropdownList` options (when they are getting dynamically from the code) can now be refreshed from the context menu.
 
 ### Removed
 - Intentionally dropped support of legacy `IMGUI` system highlighting advantages of `UI Toolkit`.